aws-db-monitoring-automation/test/requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
boto3==1.29.7
requests==2.31.0
pyyaml==6.0.1
paramiko==3.4.0
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9